Crate tree_sitter_ruby

This crate provides Ruby language support for the tree-sitter parsing library.

Typically, you will use the LANGUAGE constant to add this language to a tree-sitter Parser, and then use the parser to parse some code:

use tree_sitter::Parser;

let code = r#"
def hello(name)
 puts "Hello, #{name}!"
end
"#;
let mut parser = Parser::new();
let language = tree_sitter_ruby::LANGUAGE;
parser
    .set_language(&language.into())
    .expect("Error loading Ruby parser");
let tree = parser.parse(code, None).unwrap();
assert!(!tree.root_node().has_error());

Constants

HIGHLIGHTS_QUERY

The syntax highlighting query for this language.

LANGUAGE

The tree-sitter LanguageFn for this grammar.

LOCALS_QUERY

The local-variable syntax highlighting query for this language.

NODE_TYPES

The content of the node-types.json file for this grammar.

TAGS_QUERY

The symbol tagging query for this language.